chapters/2026/Etude-Cas-DataForGoodFrance-TeamH/assets/code/requirements.txt
pandas>=1.5.0
matplotlib>=3.7.0
pydriller>=2.9
pyyaml
numpy
tqdm>=4.0
python-dotenv>=1.0
PyGithub>=1.77
pydriller>=2.9
requests>=2.31.0
seaborn>=0.13.2